Desafio NG Billing - Desafio 02/requirements.txt
aiosmtplib>=1.2.0
aiosmtpd>=1.4.5
python-dotenv>=1.0.0
cx_Oracle>=8.3.0